+++ MLEM.Ui/Elements/ScrollBar.cs
@@ -137,15 +137,14 @@ namespace MLEM.Ui.Elements {
 
             // MOUSE INPUT
             var moused = this.Controls.MousedElement;
-            var mousedPos = Vector2.Transform(this.Input.MousePosition.ToVector2(), this.Root.InvTransform);
             if (moused == this && this.Controls.Input.IsMouseButtonPressed(MouseButton.Left)) {
                 this.isMouseHeld = true;
-                this.scrollStartOffset = mousedPos - this.ScrollerPosition;
+                this.scrollStartOffset = this.TransformInverseAll(this.Input.MousePosition.ToVector2()) - this.ScrollerPosition;
             } else if (this.isMouseHeld && !this.Controls.Input.IsMouseButtonDown(MouseButton.Left)) {
                 this.isMouseHeld = false;
             }
             if (this.isMouseHeld)
-                this.ScrollToPos(mousedPos);
+                this.ScrollToPos(this.TransformInverseAll(this.Input.MousePosition.ToVector2()));
             if (!this.Horizontal && moused != null && (moused == this.Parent || moused.GetParentTree().Contains(this.Parent))) {
                 var scroll = this.Input.LastScrollWheel - this.Input.ScrollWheel;
                 if (scroll != 0)
@@ -157,7 +156,7 @@ namespace MLEM.Ui.Elements {
                 // are we dragging on top of the panel?
                 if (this.Input.GetGesture(GestureType.VerticalDrag, out var drag)) {
                     // if the element under the drag's start position is on top of the panel, start dragging
-                    var touched = this.Parent.GetElementUnderPos(Vector2.Transform(drag.Position, this.Root.InvTransform));
+                    var touched = this.Parent.GetElementUnderPos(this.TransformInverseAll(drag.Position));
                     if (touched != null && touched != this)
                         this.isDragging = true;
 
@@ -173,7 +172,7 @@ namespace MLEM.Ui.Elements {
                 this.isTouchHeld = false;
             } else {
                 foreach (var loc in this.Input.TouchState) {
-                    var pos = Vector2.Transform(loc.Position, this.Root.InvTransform);
+                    var pos = this.TransformInverseAll(loc.Position);
                     // if we just started touching and are on top of the scroller, then we should start scrolling
                     if (this.DisplayArea.Contains(pos) && !loc.TryGetPreviousLocation(out _)) {
                         this.isTouchHeld = true;